control(feat): move as external component
